% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{addval}
\alias{addval}
\title{Title}
\usage{
addval(
  runfun,
  design,
  dat = NULL,
  points = NULL,
  each = 1,
  minrun = F,
  n.points = 10,
  seed = 1
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
