% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.simfun.R
\name{example.simfun}
\alias{example.simfun}
\title{Example simulation functions}
\usage{
example.simfun(name)
}
\arguments{
\item{name}{Name of the simulation function as a character. Currently available are: "ttest","anova","glm2","irt1","irt2","multi2".}
}
\value{
The returned object is the simulation function.
}
\description{
These simulation functions can be loaded as examples or templates. They correspond to the functions in the "simulation_functions" vignette. The vignette contains the code and a description of the simfuns.
}
\examples{
simfun = example.simfun("ttest")
simfun(400)

}
