% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions_results.R
\docType{data}
\name{extensions_results}
\alias{extensions_results}
\title{Results of the extensions vignette}
\format{
A list with designresult objects created in the vignette "extensions"
}
\usage{
extensions_results
}
\description{
Since the Monte Carlo simulations required to generate the results in the extensions vignette can be time-consuming, we have precomputed them and included them as a data file.
}
\keyword{datasets}
