% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogates.R
\name{gauss.pred}
\alias{gauss.pred}
\title{Gaussian Process Regression}
\usage{
gauss.pred(
  dat,
  goal = NULL,
  cost = function(x) sum(x),
  design = NULL,
  fixed_cost = NULL,
  greedy = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Gaussian Process Regression
}
