% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getweight}
\alias{getweight}
\title{Title}
\usage{
getweight(dat, weight.type = "freq", correct_zero = T)
}
\arguments{
\item{correct_zero}{Adds one number to the data if it consists only of 0s or only of 1s. Variance can only then be calculated.}
}
\value{

}
\description{
Title
}
