% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{power}
\alias{power}
\title{Power}
\usage{
power(
  model = model,
  altmodel = altmodel,
  test = test1,
  effect.size = 0.01,
  alpha = 0.05,
  n.pers = n.pers,
  runs = 2
)
}
\arguments{
\item{model}{model parameters created from the corresponding function}

\item{altmodel}{altmodel functions created from the corresponding function}

\item{test}{test functions created from the corresponding function}

\item{effect.size}{effect size (currently only expected Kullback-Leibler Distance)}

\item{alpha}{Alpha Niveau}

\item{n.pers}{Number of Persons}

\item{runs}{Number of Runs}
}
\value{
Power or 1-Beta
}
\description{
perform Power analysis
}
