% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogates.R
\name{reg.pred}
\alias{reg.pred}
\title{Linear Regression}
\usage{
reg.pred(
  dat,
  goal = 0.8,
  cost = function(x) sum(x),
  design = NULL,
  greedy = TRUE,
  ...
)
}
\arguments{
\item{fun}{}
}
\value{

}
\description{
Linear Regression
}
