% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{setup.alternative}
\alias{setup.alternative}
\title{Alternative Models}
\usage{
setup.alternative(class = "2PL")
}
\arguments{
\item{class}{Type of alternative model. 2PL or DIF currently}
}
\value{
Function to initialize model with given effect size, function to create dataset from model parameters
}
\description{
Specify alternative models. Instance functions have to be defined to create a model with the desired effect size difference.
}
