% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{setup.model}
\alias{setup.model}
\title{setup Models}
\usage{
setup.model(
  architecture = "Rasch",
  n.items = 20,
  alpha = c(),
  beta = c(),
  pers.dist = "normal"
)
}
\arguments{
\item{architecture}{IRT Model, e.g. Rasch}

\item{n.items}{Number of items}

\item{beta}{Vector of length n.items specifiying item difficulty}

\item{pers.dist}{Distribution of Person Parameters, e.g. normal}
}
\value{
List of model parameters and distribution function
}
\description{
Setting up the IRT Model that corresponds to the null hypothesis being tested.
}
