% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{setup.test}
\alias{setup.test}
\title{setup Tests}
\usage{
setup.test(type = "AndersenLR")
}
\arguments{
\item{type}{Test, currently: AndersenLR, M2}
}
\value{
Function that takes a dataset and returns a p-value
}
\description{
Wrapper for Tests
}
