% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_power.R
\name{sim_power}
\alias{sim_power}
\title{Title}
\usage{
sim_power(
  model = NULL,
  runfun,
  n.pers,
  max.runs = 10,
  alpha = 0.05,
  showprogress = FALSE
)
}
\arguments{
\item{alpha}{}
}
\value{

}
\description{
Basic function to estimate power at a sample size.
}
