% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_samplesize.R
\name{sim_samplesize}
\alias{sim_samplesize}
\title{Title}
\usage{
sim_samplesize(
  model = NULL,
  runfun,
  max.runs = 500,
  tol = 0,
  power = 0.8,
  guess = 500,
  alpha = 0.05,
  showprogress = FALSE
)
}
\arguments{
\item{alpha}{}
}
\value{

}
\description{
Wrapper function for different methods to approximate necessary sample size.
}
