% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_fun.R
\name{ss.find}
\alias{ss.find}
\title{Sample Size Finding Algorithm}
\usage{
ss.find(
  learner,
  runfun,
  design,
  goal = NULL,
  goal.ci = NA,
  CI = 0.95,
  budget = NULL,
  setsize = NULL,
  n.startsets = NULL,
  seed = NULL,
  dat = NULL,
  dat.only = FALSE,
  limitmaxbudget = TRUE,
  fixed_cost = NULL,
  cost = function(x) sum(x),
  ...
)
}
\arguments{
\item{setsize}{The number of draws from the data generating function in each iteration}

\item{...}{}
}
\value{

}
\description{
Sample Size Finding Algorithm
}
