% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogates.R
\name{svm.pred}
\alias{svm.pred}
\title{SVR}
\usage{
svm.pred(
  dat,
  goal = NULL,
  carryover = NULL,
  cost = function(x) sum(x),
  tune = TRUE,
  design = NULL,
  fixed_cost = NULL,
  greedy = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
SVR
}
