% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.design.R
\name{find.design}
\alias{find.design}
\title{Sample Size Finding}
\usage{
find.design(
  dgfun,
  boundaries,
  power = NULL,
  runs = 1000,
  ci = NULL,
  ci_perc = 0.95,
  time = NULL,
  costfun = NULL,
  max_cost = NULL,
  surrogate = "gpr",
  n.startsets = 4,
  init.perc = 0.1,
  setsize = 50,
  dat = NULL,
  save_dir = NULL,
  Ntry = 2
)
}
\arguments{
\item{boundaries}{Kann vector oder liste sein}

\item{n.startsets}{Startsets per Dimension}

\item{init.perc}{}

\item{setsize}{The number of draws from the data generating function in each iteration}
}
\description{
Sample Size Finding
}
