% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_fun.R
\name{gauss.pred}
\alias{gauss.pred}
\title{Title}
\usage{
gauss.pred(
  dat,
  goal = NULL,
  cost = function(x) sum(x),
  design = NULL,
  fixed_cost = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
