% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_fun.R
\name{logi.pred}
\alias{logi.pred}
\title{Title}
\usage{
logi.pred(
  dat,
  goal = 0.8,
  cost = function(x) sum(x),
  trans = TRUE,
  design = NULL,
  greedy = TRUE,
  ...
)
}
\arguments{
\item{trans}{}
}
\value{

}
\description{
Title
}
\examples{
1
}
