% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.designresult.R
\name{plot.designresult}
\alias{plot.designresult}
\title{Title}
\usage{
\method{plot}{designresult}(ds, adderrorbars = NULL, addribbon = NULL, design = NULL)
}
\arguments{
\item{ds}{}
}
\description{
Title
}
