% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_fun.R
\name{svm.pred}
\alias{svm.pred}
\title{Title}
\usage{
svm.pred(
  dat,
  goal = NULL,
  carryover = NULL,
  cost = function(x) sum(x),
  tune = TRUE,
  epsil = FALSE,
  design = NULL,
  fixed_cost = NULL,
  greedy = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
generate a prediction for the sample size at the specified power
}
